Remove all extensions when inferring deploy directory name
Our deployed "packages" most of the time are named <file>.tar.<ext>,
indicating that they are tared and compressed. In order to infer the
deploy directory name, we remove the package's extension, but that only
removes the outer most extensions (<ext>, in the above example). What we
really want is to remove all extensions instead. This change adjusts the
code accordingly.