Remove additional context from clap errors
From the looks of it, clap reported errors are basically meant to stand
on their own. In our case, the additional "failed to parse program
arguments" context we add doesn't help at all and just causes the
resulting error report to look awkward:

  > Error: failed to parse program arguments
  >
  > Caused by:
  >     error: unexpected argument '-a' found
  >
  >       tip: to pass '-a' as a value, use '-- -a'
  >
  >     Usage: chroot-deploy [OPTIONS] [MESSAGE]
  >
  >     For more information, try '--help'.

Omit this context for a slightly improved experience.